% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AreaFreedomDesign.R
\name{AreaFreedomDesign}
\alias{AreaFreedomDesign}
\alias{AreaFreedomDesign.Context}
\title{Area freedom design base class builder}
\usage{
AreaFreedomDesign(
  context,
  detected = FALSE,
  pr_detect = NULL,
  pr_persist = 1,
  iterations = NULL,
  class = character(),
  ...
)

\method{AreaFreedomDesign}{Context}(
  context,
  detected = FALSE,
  pr_detect = NULL,
  pr_persist = 1,
  iterations = NULL,
  class = character(),
  ...
)
}
\arguments{
\item{context}{A \code{Context} or inherited class object representing the
context of a bio-security surveillance and area freedom design.}

\item{detected}{A logical (> 0) temporal vector recording when the invasive
species was detected/sighted at previous intervals. Default is
\code{FALSE} implying nothing detected.}

\item{pr_detect}{The probability of detecting the invasive species given its
presence. Also known as system sensitivity or detection confidence for
surveillance systems. Default is \code{NULL} implying only detection
records are available. Temporally changing values may be provided by a
numeric vector, the length of which should be sufficient for the expected
number of \code{iterations}, given the specified stopping criteria, else
the last value of the vector is repeated.}

\item{pr_persist}{The probability that the invasive species persists at each
time interval (specified by the \code{time_unit} parameter in the
\code{context}). Default is \code{1} implies that the invasive species
will persist across time intervals if present, representing the worst case
scenario when persistence probability is unknown. Temporally changing
values may be provided by a numeric vector, the length of which should be
sufficient for the expected number of \code{iterations}, given the
specified stopping criteria, else the last value of the vector is repeated.}

\item{iterations}{The number of time intervals (specified by the
\code{time_unit} parameter in the \code{context}), or sequential
surveillance system applications, used to estimate the likelihood
of area freedom. Default is \code{NULL} implying another stopping
mechanism is utilized, such as a threshold probability or target
confidence in freedom (see inherited classes).}

\item{...}{Additional parameters.}
}
\value{
A \code{AreaFreedomDesign} base class object (list) containing
functions to provide evidence for area freedom, given detection records,
over a number of iterations, including the likelihood an invasive species
evades detection (for hypothesis testing), or the confidence in freedom
(obtained via Bayesian approaches). The following template functions are
provided in the base class:
\describe{
\item{\code{get_evidence()}}{Get a sequence of values that provide
evidence for area freedom, for each iteration or time interval
specified, or for the iterations required to satisfy the stopping
criteria (see inherited classes).}
\item{\code{get_iterations()}}{Get the number of time intervals, or
sequential surveillance system applications, used to provide evidence
of area freedom. The value returned will be either the specified
\code{iterations} parameter, or the iterations recorded when a
specified stopping criteria for the freedom evidence has been
satisfied, such as a threshold probability or target confidence in
freedom (see inherited classes).}
\item{\code{save_design()}}{Save the area freedom design as a
comma-separated value (CSV) file containing iterative evidence for
area freedom.}
}
}
\description{
Builds a base class to represent area freedom design functionality for
assessing probabilities relating to the presence or absence of an invasive
species over a sequence time intervals or applications of a surveillance
system.
}
\references{
Rout, T. (2017). Declaring Eradication of an Invasive Species. In
A. Robinson, T. Walshe, M. Burgman, & M. Nunn (Eds.),
\emph{Invasive Species: Risk Assessment and Management} (pp. 334-347).
Cambridge: Cambridge University Press.
\doi{doi.org/10.1017/9781139019606.017}
}
