% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AreaGrowthSurvDesign.R
\name{AreaGrowthSurvDesign}
\alias{AreaGrowthSurvDesign}
\alias{AreaGrowthSurvDesign.Context}
\title{Area growth surveillance design class builder}
\usage{
AreaGrowthSurvDesign(
  context,
  divisions,
  subregion_area,
  establish_rate,
  growth_rate,
  size_class_max = 10,
  class_pops_max = 100,
  f_area_growth = function(g, s) {
     pi * (g * s)^2
 },
  sample_sens = 1,
  mgmt_cost = list(),
  sample_cost = NULL,
  budget = NULL,
  ...
)

\method{AreaGrowthSurvDesign}{Context}(
  context,
  divisions,
  subregion_area,
  establish_rate,
  growth_rate,
  size_class_max = 10,
  class_pops_max = 100,
  f_area_growth = function(g, s) {
     pi * (g * s)^2
 },
  sample_sens = 1,
  mgmt_cost = list(),
  sample_cost = NULL,
  budget = NULL,
  ...
)
}
\arguments{
\item{context}{A \code{Context} or inherited class object representing the
context of a bio-security surveillance and area freedom design.}

\item{divisions}{A \code{Divisions} or inherited class object representing
one or more sub-regions for the surveillance design.}

\item{subregion_area}{A vector of values to represent the area covered by
each sub-region specified by \code{divisions}. Units should be consistent
with the \code{dist_unit} parameter specified in the \code{context}.}

\item{establish_rate}{A vector of values to represent to the expected number
of populations to establish in each sub-region (specified by
\code{divisions}) within one time interval specified by the
\code{time_unit} parameter in the \code{context}.}

\item{growth_rate}{The area-based radial population growth rate in distance
per time interval in units specified by the \code{dist_unit} and
\code{time_unit} parameters in the \code{context}.}

\item{size_class_max}{The maximum size class for populations represented in
the growth model. Default is \code{10}}

\item{class_pops_max}{The maximum number of populations in a size class.
Default is \code{100}}

\item{f_area_growth}{A function for calculating the area of a population
given its size class and the \code{growth_rate}. The function should be in
the form \code{function(growth_rate, size_class)}. The default function
implements radial (circular) expansion:
\code{pi*(growth_rate*size_class)^2}.}

\item{sample_sens}{A vector of sample sensitivity values for each sub-region
specified by \code{divisions}. Default is \code{1}.}

\item{mgmt_cost}{A list of vectors to represent estimated management costs
for population incursions. Each vector specifies costs at each sub-region
specified by \code{divisions}. Named list elements should include
\code{eradication} cost per area,  \code{damage} cost per area, and
\code{penalty} cost per population of maximum size class. Default is an
empty list. Units should be consistent with the \code{cost_unit} and
\code{dist_unit} (area) parameters specified in the \code{context}.}

\item{sample_cost}{A vector of cost per sample of allocated surveillance
resources at each sub-region specified by \code{divisions}. Default is
\code{NULL}. Units should be consistent with the \code{cost_unit}
parameter specified in the \code{context}.}

\item{budget}{The cost budget or constraint for the sampling allocation in
the surveillance design. Default is \code{NULL}. Units should be
consistent with \code{sample_cost} when specified. Otherwise the units
should be consistent with the \code{surv_qty_unit} parameter specified in
the \code{context} (e.g. traps or samples).}

\item{...}{Additional parameters.}
}
\value{
A \code{AreaGrowthSurvDesign} class object (list) containing inherited
and extended functions from the base \code{SurveillanceDesign} class for
for allocating sample densities, and calculating (unit and overall)
detection sensitivities:
\describe{
\item{\code{get_allocation()}}{Get allocated sample densities via
specified strategy, utilizing costs and budget constraints.}
\item{\code{get_sensitivity()}}{Get the sub-region detection
sensitivities of the allocated surveillance design.}
\item{\code{get_confidence(growth = NULL)}}{Get the overall system
sensitivity or detection confidence of the allocated surveillance
design. The optional \code{growth} parameter may provide a vector of
relative increasing multipliers (e.g. 1, 1.8, 4.3, 7.5) applied to the
sampling design density implied by the method over time or a sequence
of repeated surveillance efforts, which provide a proxy for invasive
species growth. When present, increasing system sensitivity values are
returned for each multiplier or time/repeat.}
\item{\code{set_cores(cores)}}{Set the number of cores available for
parallel processing and thus enable parallel processing for
calculating optimal sample density allocation.}
\item{\code{set_precision(prec)}}{Set the precision used in the search
for budget constrained allocation. Default is 3. Higher values may
result in excessive computational times.}
}
}
\description{
Builds a class to represent surveillance design functionality for the
cost-effective allocation of surveillance sampling densities across one or
more spatial sub-regions via a method that incorporates area-based
population growth, surveillance and management costs, including costs
associated with eradication, damages, and failure to contain incursions.
}
\references{
Epanchin-Niell, R. S., Haight, R. G., Berec, L., Kean, J. M., & Liebhold,
A. M. (2012). Optimal surveillance and eradication of invasive species in
heterogeneous landscapes. \emph{Ecology Letters}, 15(8), 803–812.
\doi{10.1111/j.1461-0248.2012.01800.x}
}
