% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HypothTestFreedomDesign.R
\name{HypothTestFreedomDesign}
\alias{HypothTestFreedomDesign}
\alias{HypothTestFreedomDesign.Context}
\title{Hypothesis testing area freedom design base class builder}
\usage{
HypothTestFreedomDesign(
  context,
  detected = FALSE,
  pr_detect = NULL,
  pr_persist = 1,
  iterations = NULL,
  p_value = NULL,
  ...
)

\method{HypothTestFreedomDesign}{Context}(
  context,
  detected = FALSE,
  pr_detect = NULL,
  pr_persist = 1,
  iterations = NULL,
  p_value = NULL,
  ...
)
}
\arguments{
\item{context}{A \code{Context} or inherited class object representing the
context of a bio-security surveillance and area freedom design.}

\item{detected}{A logical (> 0) temporal vector recording when the invasive
species was detected/sighted at previous intervals. Default is
\code{FALSE} implying nothing detected.}

\item{pr_detect}{The probability of detecting the invasive species given its
presence. Also known as system sensitivity or detection confidence for
a surveillance system. Default is \code{NULL} implying only detection
records are available.}

\item{pr_persist}{The probability that the invasive species persists at each
time interval (specified by the \code{time_unit} parameter in the
\code{context}). Default is \code{1} implies that the invasive species
will persist across time intervals if present, representing the worst case
scenario when persistence probability is unknown.}

\item{iterations}{The number of time intervals (specified by the
\code{time_unit} parameter in the \code{context}), or sequential
surveillance system applications, used to estimate the likelihood of area
freedom. Default is \code{NULL} implying that the \code{p_value} parameter
will be utilized as a stopping mechanism.}

\item{p_value}{The threshold probability (e.g. 0.05) for rejecting the null
hypothesis that the invasive species remains present given a sequence of
no detection via a surveillance system. Default is \code{NULL} implying
that the \code{iterations} parameter will be utilized as a stopping
mechanism.}

\item{...}{Additional parameters.}
}
\value{
A \code{HypothTestFreedomDesign} class object (list) containing
inherited and extended functions from the base \code{AreaFreedomDesign}
class to provide evidence for area freedom, given detection records,
over a number of iterations, via hypothesis testing approaches:
\describe{
\item{\code{get_evidence()}}{Get a sequence of values for the
probability that an extant invasive species could evade surveillance
system detection, for each iteration or time interval specified, or
for the iterations required to satisfy the threshold probability
\code{p_value} for rejecting the null hypothesis that the invasive
species remains present.}
\item{\code{get_iterations()}}{Get the number of time intervals, or
sequential surveillance system applications, used to provide evidence
for area freedom. The value returned will be either the specified
\code{iterations} parameter, or the iterations recorded when the
specified threshold probability \code{p_value} for rejecting the null
hypothesis of presence given no detection is reached.}
}
}
\description{
Builds a class to represent area freedom design functionality utilizing
hypothesis testing approaches to assess the likelihood of an invasive
species being present when it has not been detected for a sequence of time
intervals or applications of a surveillance system.
}
\references{
Barclay, H. J., & Hargrove, J. W. (2005). Probability models to facilitate
a declaration of pest-free status, with special reference to tsetse
(Diptera: Glossinidae). \emph{Bulletin Of Entomological Research}
\emph{(London)}, 95(1), 1–12. \doi{10.1079/BER2004331}

Regan, T. J., McCarthy, M. A., Baxter, P. W., Dane Panetta, F., &
Possingham, H. P. (2006). Optimal eradication: when to stop looking for
an invasive plant. \emph{Ecology Letters}, 9(7), 759–766.
\doi{10.1111/j.1461-0248.2006.00920.x}

Rout, T. (2017). Declaring Eradication of an Invasive Species. In
A. Robinson, T. Walshe, M. Burgman, & M. Nunn (Eds.),
\emph{Invasive Species: Risk Assessment and Management} (pp. 334-347).
Cambridge: Cambridge University Press.
\doi{doi.org/10.1017/9781139019606.017}

Solow, A. R. (1993). Inferring Extinction from Sighting Data.
\emph{Ecology}, 74(3), 962–964. https://doi.org/10.2307/1940821
}
