% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultilevelSurvDesign.R
\name{MultilevelSurvDesign}
\alias{MultilevelSurvDesign}
\alias{MultilevelSurvDesign.Context}
\title{Multilevel sampling surveillance design class builder}
\usage{
MultilevelSurvDesign(
  context,
  divisions,
  sample_sens = 1,
  sample_type = c("discrete", "continuous"),
  prevalence = NULL,
  total_indiv = NULL,
  design_dens = NULL,
  sample_area = NULL,
  sample_cost = 1,
  system_sens = 0.95,
  class = character(),
  ...
)

\method{MultilevelSurvDesign}{Context}(
  context,
  divisions,
  sample_sens = 1,
  sample_type = c("discrete", "continuous"),
  prevalence = NULL,
  total_indiv = NULL,
  design_dens = NULL,
  sample_area = NULL,
  sample_cost = 1,
  system_sens = 0.95,
  class = character(),
  ...
)
}
\arguments{
\item{context}{A \code{Context} or inherited class object representing the
context of a bio-security surveillance and area freedom design.}

\item{divisions}{A \code{Divisions} or inherited class object representing
one or more multistage levels, for the surveillance design. The levels
should be ordered from lowest to highest (e.g. leaves, trees, rows,
orchards). Only the lowest level can utilize either continuous
density-based or discrete sampling, higher levels can only utilize
discrete sampling. Note that more than five levels may not be
computationally feasible.}

\item{sample_sens}{A single sample sensitivity value for the lowest level
of a multilevel/stage sampling. Default is \code{1}.}

\item{sample_type}{The type of sampling used for the lowest level of the
multilevel sampling. One of \code{"discrete"} or \code{"continuous"}.}

\item{prevalence}{A vector of discrete sampling design prevalence values for
each multistage level specified by \code{divisions}. Note that this
parameter may represent apparent prevalence (Cannon, 2009) when the
\code{sensitivity} is set to \code{1}. Default is \code{NULL}.}

\item{total_indiv}{A vector of total individual discrete sampling units
(e.g. leaves, trees, rows, orchards) present at each multistage level
specified by \code{divisions}. Default is \code{NULL}.}

\item{design_dens}{A single design density value for when the lowest level
of the multilevel sampling is continuous. Default is \code{NULL}.}

\item{sample_area}{The area of a single sample when the lowest level for
when the lowest level of the multilevel sampling is continuous. Note that
when set to 1, the total number of samples will be equivalent to the total
area sampled. Default is \code{NULL}.}

\item{sample_cost}{The cost of samples at each level. Default is \code{1}.
An attribute \code{units} may be used to specify the cost units (e.g. "$"
or "hours").}

\item{system_sens}{The desired (minimum) system sensitivity or detection
probability of the surveillance design (e.g. 0.95). Default is
\code{NULL}.}

\item{...}{Additional parameters.}
}
\value{
A \code{MultilevelSurvDesign} class object (list) containing
inherited and extended functions from the base \code{SurveillanceDesign}
class for for allocating resources, and calculating (unit and overall)
detection sensitivities:
\describe{
\item{\code{get_context()}}{Get context object.}
\item{\code{get_divisions()}}{Get divisions object.}
\item{\code{get_allocation()}}{Get allocated surveillance resources for
each level to achieve a minimal total cost given a specified desired
system sensitivity or detection probability.}
\item{\code{get_sensitivity()}}{Get the level detection
sensitivities of the allocated surveillance design.}
\item{\code{get_system_sens(growth = NULL)}}{Get the overall system
sensitivity or detection probability of the allocated surveillance
design. The optional \code{growth} parameter may provide a vector of
relative increasing multipliers (e.g. 1, 1.8, 4.3, 7.5) applied to the
discrete sampling design \code{prevalence} over time or a sequence of
repeated surveillance efforts, which provide a proxy for invasive
species growth. When present, increasing system sensitivity values are
returned for each multiplier or time/repeat.}
\item{\code{save_design()}}{Save the surveillance design as a
collection of comma-separated value (CSV) files, including the
surveillance \code{allocation}, \code{sensitivity}, and a
\code{summary} of sample costs (when applicable) and the system
sensitivity (detection probability).}
}
}
\description{
Builds a class to represent surveillance design functionality for the
effective allocation of surveillance sampling across one or more levels or
stages via multilevel sampling methods specified with surveillance detection
sensitivities, sample costs, and overall detection probability.
}
\references{
Cannon, R. M. (2009). Inspecting and monitoring on a restricted budget -
where best to look? \emph{Preventive Veterinary Medicine}, 92(1–2),
163-174. \doi{10.1016/j.prevetmed.2009.06.009}

Kean, J. M., Burnip, G. M. & Pathan, A. (2015). Detection survey design
for decision making during biosecurity incursions. In F. Jarrad,
S. Low-Choy & K. Mengersen (eds.),
\emph{Biosecurity surveillance: Quantitative approaches} (pp. 238– 250).
Wallingford, UK: CABI.
}
