% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RangeKernelSurvDesign.R
\name{RangeKernelSurvDesign}
\alias{RangeKernelSurvDesign}
\alias{RangeKernelSurvDesign.Context}
\title{Range kernel surveillance design class builder}
\usage{
RangeKernelSurvDesign(
  context,
  divisions,
  establish_pr,
  lambda,
  sigma,
  intervals,
  prevalence = 1,
  optimal = c("detection", "none"),
  budget = NULL,
  confidence = NULL,
  exist_alloc = NULL,
  exist_sens = NULL,
  ...
)

\method{RangeKernelSurvDesign}{Context}(
  context,
  divisions,
  establish_pr,
  lambda,
  sigma,
  intervals,
  prevalence = 1,
  optimal = c("detection", "none"),
  budget = NULL,
  confidence = NULL,
  exist_alloc = NULL,
  exist_sens = NULL,
  ...
)
}
\arguments{
\item{context}{A \code{Context} or inherited class object representing the
context of a bio-security surveillance and area freedom design.}

\item{divisions}{A \code{Divisions} or inherited class object representing
spatial locations (grid-cells) for the surveillance design.}

\item{establish_pr}{A vector of (relative) probability values to represent
the likelihood of pest establishment at each spatial location specified
by \code{divisions}. Values are assumed to be relative when their
maximum is greater than 1, or an attribute \code{relative = TRUE} is
attached to the parameter.}

\item{lambda}{A vector of maximum detection or capture rates for each
spatial location specified by \code{divisions}. Maximum detection/capture
is achieved when a surveillance resource is placed in the center of the
home range of an invasive species individual or colony. Detection/capture
rate decays with distance via an exponential kernel such that the
detection/capture probability of each allocated resource can be expressed
via \code{1 - (1 - lambda*exp(-distance^2/(2*sigma^2))^intervals)} for a
given number of time \code{intervals} and up to a maximum distance of
\code{4*sigma}, where \code{sigma} is the home range decay parameter.}

\item{sigma}{The spatial decay parameter for the (half-Normal) home range
kernel (see \code{lambda}). Note that \code{sigma} also specifies the
maximum effective distance of detection/capture resources (i.e. twice the
home range radius or \code{4*sigma}). Units should be consistent
with the \code{dist_unit} parameter specified in the \code{context}.}

\item{intervals}{The number of time intervals each detection/capture
resource or device is utilized (e.g. nights a trap is set).}

\item{prevalence}{The cell-level design prevalence indicating the minimum
number of location cells that are expected to be infected with the
invasive species if the region of interest specified by \code{divisions}
is infected. Default is \code{1}. Higher values can be used as a proxy
to population growth over time.}

\item{optimal}{The strategy used for finding an effective surveillance
resource allocation. Either (maximum) \code{"detection"} sensitivity (up
to \code{"confidence"} level when specified), or \code{"none"} for
representing existing surveillance designs only.}

\item{budget}{The budget or constraint for the number of surveillance
resource units or devices available for allocation in the surveillance
design. Default is \code{NULL}.}

\item{confidence}{The desired (minimum) system sensitivity or detection
confidence of the surveillance design (e.g. 0.95). Default is \code{NULL}.}

\item{exist_alloc}{A vector of existing surveillance resource quantities at
each spatial location specified by \code{divisions}, or a data frame
(or matrix) of resource location coordinates in longitude and latitude
(WGS84) with explicitly named columns "lon" and "lat". Should only be used
to represent existing surveillance designs when \code{optimal = "none"}.
Default is \code{NULL}.}

\item{exist_sens}{A vector, or list of vectors, of detection sensitivity
values of existing surveillance present at each spatial location specified
by \code{divisions}. Multiple existing surveillance layers may be
specified in a list. Default is \code{NULL}.}

\item{...}{Additional parameters.}
}
\value{
A \code{RangeKernelSurvDesign} class object (list) containing
inherited and extended functions from the base \code{SurveillanceDesign}
class for for allocating resources, and calculating (location and overall)
detection sensitivities:
\describe{
\item{\code{get_allocation(coords = FALSE)}}{Get allocated resources via
specified strategy, utilizing costs, benefits, budget constraints,
and/or desired detection confidence level. Returns resources at each
location when \code{coords = FALSE} (default), or resource location
coordinates in longitude and latitude (WGS84) when
\code{coords = TRUE}.}
\item{\code{get_sensitivity()}}{Get the location detection sensitivities
of the allocated surveillance design.}
\item{\code{get_confidence(growth = NULL)}}{Get the overall system
sensitivity or detection confidence of the allocated surveillance
design. The optional \code{growth} parameter may provide a vector of
relative increasing multipliers (e.g. 1, 1.8, 4.3, 7.5) applied to the
cell-level design \code{prevalence} over time or a sequence of
repeated surveillance efforts, which provide a proxy for invasive
species growth. When present, increasing system sensitivity values are
returned for each multiplier or time/repeat.}
\item{\code{set_cores(cores)}}{Set the number of cores available for
parallel processing and thus enable parallel processing for
calculating optimal resource allocation.}
}
}
\description{
Builds a class to represent surveillance design functionality for the
effective allocation of surveillance resources that are each decreasingly
effective across a spatial range, modeled via an exponential decay kernel,
and specified with a surveillance resource budget and/or an overall desired
detection confidence.
}
\references{
Anderson, D. P., Ramsey, D. S. L., Nugent, G., Bosson, M., Livingstone,
P., Martin, P. A. J., Sergeant, E., Gormley, A. M., & Warburton, B.
(2013). A novel approach to assess the probability of disease eradication
from a wild-animal reservoir host. \emph{Epidemiology and Infection},
141(7), 1509–1521. \doi{10.1017/S095026881200310X}

Anderson, D. P., Gormley, A. M., Ramsey, D. S. L., Nugent, G., Martin,
P. A. J., Bosson, M., Livingstone, P., & Byrom, A. E. (2017).
Bio-economic optimisation of surveillance to confirm broadscale
eradications of invasive pests and diseases. \emph{Biological Invasions},
19(10), 2869–2884. \doi{10.1007/s10530-017-1490-5}

Anderson, D. P., Pepper, M. A., Travers, S., Michaels, T. A., Sullivan,
K., & Ramsey, D. S. L. (2022). Confirming the broadscale eradication
success of nutria (Myocastor coypus) from the Delmarva Peninsula, USA.
\emph{Biological Invasions}. \doi{10.1007/s10530-022-02855-x}

McCarthy, M. A., Thompson, C. J., Hauser, C., Burgman, M. A., Possingham,
H. P., Moir, M. L., Tiensin, T., & Gilbert, M. (2010). Resource allocation
for efficient environmental management. \emph{Ecology Letters}, 13(10),
1280–1289. \doi{10.1111/j.1461-0248.2010.01522.x}

Moore, A. L., McCarthy, M. A., & Lecomte, N. (2016). Optimizing ecological
survey effort over space and time.
\emph{Methods in Ecology and Evolution}, 7(8), 891–899.
\doi{10.1111/2041-210X.12564}
}
