% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurveillanceDesign.R
\name{SurveillanceDesign}
\alias{SurveillanceDesign}
\alias{SurveillanceDesign.Context}
\title{Surveillance design base class builder}
\usage{
SurveillanceDesign(
  context,
  divisions,
  establish_pr = NULL,
  optimal = c("cost", "saving", "benefit", "detections", "sensitivity", "none"),
  mgmt_cost = list(),
  benefit = NULL,
  alloc_cost = NULL,
  fixed_cost = NULL,
  budget = NULL,
  system_sens = NULL,
  min_alloc = NULL,
  max_alloc = NULL,
  discrete_alloc = FALSE,
  exist_alloc = NULL,
  exist_sens = NULL,
  class = character(),
  ...
)

\method{SurveillanceDesign}{Context}(
  context,
  divisions,
  establish_pr = NULL,
  optimal = c("cost", "saving", "benefit", "detections", "sensitivity", "none"),
  mgmt_cost = list(),
  benefit = NULL,
  alloc_cost = NULL,
  fixed_cost = NULL,
  budget = NULL,
  system_sens = NULL,
  min_alloc = NULL,
  max_alloc = NULL,
  discrete_alloc = FALSE,
  exist_alloc = NULL,
  exist_sens = NULL,
  class = character(),
  ...
)
}
\arguments{
\item{context}{A \code{Context} or inherited class object representing the
context of a bio-security surveillance and area freedom design.}

\item{divisions}{A \code{Divisions} or inherited class object representing
one or more divisions (parts, locations, categories, etc.) for the
surveillance design.}

\item{establish_pr}{A vector of (relative) probability values to represent
the likelihood of pest establishment at each division part (location,
category, etc.) specified by \code{divisions}. Default is \code{NULL}.
Values are assumed to be relative when their maximum is greater than 1, or
an attribute \code{relative = TRUE} is attached to the parameter.}

\item{optimal}{The strategy used for finding an effective surveillance
resource allocation. One of (minimum) \code{"cost"}, (maximum)
\code{"saving"} (or cost-dependent benefit), (maximum) \code{"benefit"}
(independent of surveillance costs), (maximum) number of
\code{"detections"}, (maximum) overall system-wide \code{"sensitivity"}
or \code{"none"} for representing existing surveillance designs only.}

\item{mgmt_cost}{A list of vectors to represent management costs specific to
the method implemented in the inherited class. Each vector specifies costs
at each division part (location, category, etc.) specified by
\code{divisions}. Default is an empty list. Units should be consistent
with the \code{cost_unit} parameter specified in the \code{context}.}

\item{benefit}{A vector of values quantifying the benefit (or cost-based
saving) associated with detection at each division part (location,
category, etc.) specified by \code{divisions}. Default is \code{NULL}.
When the benefit refers to cost-based savings (i.e. \code{optimal} is
\code{"saving"}), then the units should be consistent with the
\code{cost_unit} parameter specified in the \code{context}.}

\item{alloc_cost}{A vector of cost per unit of allocated surveillance
resources. Default is \code{NULL}. Units should be consistent with the
\code{cost_unit} parameter specified in the \code{context}.}

\item{fixed_cost}{A vector of fixed costs, such as travel costs or time, at
each division part (location, category, etc.) specified by
\code{divisions}. Default is \code{NULL}. Units should be consistent with
\code{alloc_cost} when specified. Otherwise the units should be consistent
with the \code{surv_qty_unit} parameter specified in the \code{context}.}

\item{budget}{The cost budget or constraint for the resource allocation in
the surveillance design. Default is \code{NULL}. Units should be
consistent with \code{alloc_cost} when specified. Otherwise the units
should be consistent with the \code{surv_qty_unit} parameter specified in
the \code{context}.}

\item{system_sens}{The desired (minimum) system sensitivity or detection
probability of the surveillance design (e.g. 0.95). Default is
\code{NULL}.}

\item{min_alloc}{A vector of minimum permissible allocated surveillance
resource quantities at each division part (location, category, etc.)
specified by \code{divisions}. Used to avoid impractically low allocation
quantities. Default is \code{NULL}.}

\item{max_alloc}{A vector of maximum permissible allocated surveillance
resource quantities at each division part (location, category, etc.)
specified by \code{divisions}. Used to avoid impractically high allocation
quantities. Default is \code{NULL}.}

\item{discrete_alloc}{A logical to indicate that the allocated surveillance
resource quantities at each division part (location, category, etc.)
specified by \code{divisions} should be discrete integers. Used to
allocate discrete surveillance units, such as traps or detectors. Default
is \code{FALSE} for continuous resources quantities, such as survey hours.}

\item{exist_alloc}{A vector or matrix (containing temporal columns) of
existing surveillance resource quantities at each division part (location,
category, etc.) (rows) specified by \code{divisions}. Should only be used
to represent existing surveillance designs when \code{optimal = "none"}.
Default is \code{NULL}.}

\item{exist_sens}{A vector, or list of vectors, of detection sensitivity
values of existing surveillance present at each division part (location,
category, etc.) specified by \code{divisions}. Multiple existing
surveillance layers may be specified in a list. Default is \code{NULL}.}

\item{...}{Additional parameters.}
}
\value{
A \code{SurveillanceDesign} class object (list) containing functions
for allocating resources, and calculating (unit and overall) detection
sensitivities:
\describe{
\item{\code{get_context()}}{Get context object.}
\item{\code{get_divisions()}}{Get divisions object.}
\item{\code{get_allocation()}}{Get allocated resources via specified
strategy, utilizing costs, benefits, budget constraints, and/or
desired system sensitivity or detection probability.}
\item{\code{get_sensitivity()}}{Get the division part detection
sensitivities of the allocated surveillance design combined with any
existing sensitivities specified via \code{exist_sens}.}
\item{\code{get_system_sens(growth = NULL)}}{Get the overall system
sensitivity or detection probability of the surveillance design. The
optional \code{growth} parameter may provide a vector of relative
increasing multipliers (e.g. 1, 1.8, 4.3, 7.5) applied to the
prevalence or density of the design over time or a sequence of
repeated surveillance efforts, which provide a proxy for invasive
species growth. When present, increasing system sensitivity values are
returned for each multiplier or time/repeat.}
\item{\code{save_design(...)}}{Save the surveillance design as a
collection of raster TIF and/or comma-separated value (CSV) files,
appropriate for the \code{divisions} type, including the surveillance
\code{allocation}, the corresponding sensitivity \code{alloc_sens}
when existing sensitivity is included, the overall \code{sensitivity},
\code{surv_cost} (combined surveillance allocation and fixed costs),
and a \code{summary} (CSV) of the total allocation, total costs (when
applicable), and the overall allocation sensitivity
(\code{alloc_sens}) when applicable, and the system-wide sensitivity
(\code{system_sens}). \code{Terra} raster write options may be passed
to the function for saving grid-based designs.}
}
}
\description{
Builds a base class to represent surveillance design functionality for the
effective allocation of surveillance resources across one or more divisions
(parts, locations, categories, etc.) via methods that utilize surveillance
and/or incursion management costs, benefits, detection sensitivities,
and/or an overall desired system-wide sensitivity or detection probability.
}
\references{
Cannon, R. M. (2009). Inspecting and monitoring on a restricted budget -
where best to look? \emph{Preventive Veterinary Medicine}, 92(1–2),
163-174. \doi{10.1016/j.prevetmed.2009.06.009}
}
