% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LagrangeSurvDesign.R
\name{LagrangeSurvDesign}
\alias{LagrangeSurvDesign}
\alias{LagrangeSurvDesign.Context}
\title{Lagrange surveillance design class builder}
\usage{
LagrangeSurvDesign(
  context,
  divisions,
  establish_pr,
  lambda,
  optimal = c("cost", "benefit", "detection"),
  mgmt_cost = NULL,
  benefit = NULL,
  alloc_units = NULL,
  fixed_cost = NULL,
  budget = NULL,
  confidence = NULL,
  exist_sens = NULL,
  class = character(),
  ...
)

\method{LagrangeSurvDesign}{Context}(
  context,
  divisions,
  establish_pr,
  lambda,
  optimal = c("cost", "benefit", "detection"),
  mgmt_cost = NULL,
  benefit = NULL,
  alloc_units = NULL,
  fixed_cost = NULL,
  budget = NULL,
  confidence = NULL,
  exist_sens = NULL,
  class = character(),
  ...
)
}
\arguments{
\item{context}{A \code{Context} or inherited class object representing the
context of a bio-security surveillance and area freedom design.}

\item{divisions}{A \code{Divisions} or inherited class object representing
one or more divisions (parts, locations, categories, etc.) for the
surveillance design.}

\item{establish_pr}{A vector of (relative) probability values to represent
the likelihood of pest establishment at each division part (location,
category, etc.) specified by \code{divisions}. Values are assumed to be
relative when their maximum is greater than 1, or an attribute
\code{relative = TRUE} is attached to the parameter.}

\item{lambda}{A vector of efficacy or detection rates for each division part
(location, category, etc.) specified by \code{divisions}, such that the
probability of detecting an incursion when present at a part can be
expressed via \code{pr(detect|presence) = 1 - exp(-lambda*allocation)},
for a given allocation of surveillance resources.}

\item{optimal}{The strategy used for finding an effective surveillance
resource allocation. One of (minimum) \code{"cost"}, (maximum)
\code{"benefit"}, or (maximum) \code{"detection"} sensitivity (up to
\code{"confidence"} level when specified).}

\item{mgmt_cost}{A list of vectors to represent management costs specific to
the method implemented in the inherited class. Each vector specifies costs
at each division part (location, category, etc.) specified by
\code{divisions}. Default is an empty list. An attribute \code{units} may
be used to specify the cost units (e.g. "$" or "hours").}

\item{benefit}{A vector of values quantifying the benefit of detection
at each division part (location, category, etc.) specified by
\code{divisions}. Default is \code{NULL}. An attribute \code{units} may
be used to specify the benefit units (e.g. "$" or "hours").}

\item{alloc_units}{The units for the allocated surveillance resource costs
(e.g. "$", "hours") consistent with the \code{context}. This may be
different to those specified for \code{mgmt_cost} or \code{benefit}.
Default is \code{NULL}.}

\item{fixed_cost}{A vector of fixed costs, such as travel costs or time, at
each division part (location, category, etc.) specified by
\code{divisions}. Default is \code{NULL}. Units are specified in
\code{alloc_units}.}

\item{budget}{The cost budget or constraint for the resource allocation in
the surveillance design. Default is \code{NULL}. Units are specified in
\code{alloc_units}.}

\item{confidence}{The desired (minimum) system detection sensitivity or
confidence of the surveillance design (e.g. 0.95). Default is \code{NULL}.}

\item{exist_sens}{A vector of detection sensitivity values of existing
surveillance present at each division part (location, category,
etc.) specified by \code{divisions}. Default is \code{NULL}.}

\item{...}{Additional parameters.}
}
\value{
A \code{LagrangeSurvDesign} class object (list) containing inherited
and extended functions from the base \code{SurveillanceDesign} class for
for allocating resources, and calculating (unit and overall) detection
sensitivities:
\describe{
\item{\code{get_allocation()}}{Get allocated surveillance resources via
specified strategy, utilizing costs, benefits, budget constraints,
and/or desired confidence level.}
\item{\code{get_sensitivity()}}{Get the division part detection
sensitivities of the allocated surveillance design.}
\item{\code{get_confidence()}}{Get the overall system sensitivity or
confidence of the allocated surveillance design.}
}
}
\description{
Builds a generic class to represent a surveillance design functionality for
the effective allocation of surveillance resources across one or more
divisions (parts, locations, categories, etc.) via Lagrange-based methods
for optimizing objective functions specified with surveillance and/or
incursion management costs, benefits, detection sensitivities, and/or
overall detection confidence.
}
\references{
Cannon, R. M. (2009). Inspecting and monitoring on a restricted budget -
where best to look? \emph{Preventive Veterinary Medicine}, 92(1–2),
163-174. \doi{10.1016/j.prevetmed.2009.06.009}

Hauser, C. E., & McCarthy, M. A. (2009). Streamlining 'search and
destroy': cost-effective surveillance for invasive species management.
\emph{Ecology Letters}, 12(7), 683–692.
\doi{10.1111/j.1461-0248.2009.01323.x}

McCarthy, M. A., Thompson, C. J., Hauser, C., Burgman, M. A., Possingham,
H. P., Moir, M. L., Tiensin, T., & Gilbert, M. (2010). Resource allocation
for efficient environmental management. \emph{Ecology Letters}, 13(10),
1280–1289. \doi{10.1111/j.1461-0248.2010.01522.x}

Moore, A. L., McCarthy, M. A., & Lecomte, N. (2016). Optimizing ecological
survey effort over space and time.
\emph{Methods in Ecology and Evolution}, 7(8), 891–899.
\doi{10.1111/2041-210X.12564}
}
