% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SamplingSurvDesign.R
\name{SamplingSurvDesign}
\alias{SamplingSurvDesign}
\alias{SamplingSurvDesign.Context}
\title{Sampling surveillance design class builder}
\usage{
SamplingSurvDesign(
  context,
  divisions,
  establish_pr,
  sample_sens = 1,
  sample_type = c("discrete", "continuous"),
  prevalence = NULL,
  total_indiv = NULL,
  design_dens = NULL,
  sample_area = NULL,
  optimal = c("cost", "saving", "benefit", "detections", "sensitivity", "none"),
  mgmt_cost = list(),
  benefit = NULL,
  sample_cost = NULL,
  fixed_cost = NULL,
  budget = NULL,
  system_sens = NULL,
  min_alloc = NULL,
  max_alloc = NULL,
  discrete_alloc = TRUE,
  exist_alloc = NULL,
  exist_sens = NULL,
  class = character(),
  ...
)

\method{SamplingSurvDesign}{Context}(
  context,
  divisions,
  establish_pr,
  sample_sens = 1,
  sample_type = c("discrete", "continuous"),
  prevalence = NULL,
  total_indiv = NULL,
  design_dens = NULL,
  sample_area = NULL,
  optimal = c("cost", "saving", "benefit", "detections", "sensitivity", "none"),
  mgmt_cost = list(),
  benefit = NULL,
  sample_cost = NULL,
  fixed_cost = NULL,
  budget = NULL,
  system_sens = NULL,
  min_alloc = NULL,
  max_alloc = NULL,
  discrete_alloc = TRUE,
  exist_alloc = NULL,
  exist_sens = NULL,
  class = character(),
  ...
)
}
\arguments{
\item{context}{A \code{Context} or inherited class object representing the
context of a bio-security surveillance and area freedom design.}

\item{divisions}{A \code{Divisions} or inherited class object representing
one or more divisions (parts, locations, categories, etc.) for the
surveillance design.}

\item{establish_pr}{A vector of (relative) probability values to represent
the likelihood of pest establishment at each division part (location,
category, etc.) specified by \code{divisions}. Values are assumed to be
relative when their maximum is greater than 1, or an attribute
\code{relative = TRUE} is attached to the parameter.}

\item{sample_sens}{A vector of sample sensitivity values for each division
part (location, category, etc.) specified by \code{divisions}. Default
is \code{1}.}

\item{sample_type}{The type of sampling used. One of \code{"discrete"} or
\code{"continuous"}.}

\item{prevalence}{A vector of discrete sampling design prevalence values for
each division part (location, category, etc.) specified by
\code{divisions}. Note that this parameter may represent apparent
prevalence (Cannon, 2009) when the \code{sample_sens} is set to \code{1}.
Default is \code{NULL}.}

\item{total_indiv}{A vector of total individual discrete sampling units
(e.g. plants, animals) present at each division part (location, category,
etc.) specified by \code{divisions}. Required when > 10\% of the total
individuals are expected to be sampled. Default \code{NULL} (unknown or
not required) assumes <= 10\% of the total individuals are sampled.}

\item{design_dens}{A vector of continuous sampling design density values for
each division part (location, category, etc.). Default is \code{NULL}.}

\item{sample_area}{The area of a single sample in a continuous sampling
design. Note that when set to \code{1}, the total number of samples will
be equivalent to the total area sampled. Default is \code{NULL}.}

\item{optimal}{The strategy used for finding an effective surveillance
resource allocation. One of (minimum) \code{"cost"}, (maximum)
\code{"saving"} (or cost-dependent benefit), (maximum) \code{"benefit"}
(independent of surveillance costs), (maximum) number of
\code{"detections"}, (maximum) overall system-wide \code{"sensitivity"}
or \code{"none"} for representing existing surveillance designs only.}

\item{mgmt_cost}{A list of vectors to represent estimated management costs
for when the incursion is detected and undetected. Each vector specifies
costs at each division part (location, category, etc.) specified by
\code{divisions}. List elements should be named \code{detected} and
\code{undetected}. Default is an empty list. Units should be consistent
with the \code{cost_unit} parameter specified in the \code{context}.}

\item{benefit}{A vector of values quantifying the benefit (or cost-based
saving) associated with detection at each division part (location,
category, etc.) specified by \code{divisions}. Default is \code{NULL}.
When the benefit refers to cost-based savings (i.e. \code{optimal} is
\code{"saving"}), then the units should be consistent with the
\code{cost_unit} parameter specified in the \code{context}.}

\item{sample_cost}{A vector of cost per sample of allocated surveillance
resources at each division part (location, category, etc.) specified by
\code{divisions}. Default is \code{NULL}. Units should be consistent with
the \code{cost_unit} parameter specified in the \code{context}.}

\item{fixed_cost}{A vector of fixed costs, such as travel costs or time, at
each division part (location, category, etc.) specified by
\code{divisions}. Default is \code{NULL}. Units should be consistent with
\code{sample_cost} when specified. Otherwise the units should be
consistent with the \code{surv_qty_unit} parameter specified in the
\code{context} (e.g. traps or samples).}

\item{budget}{The cost budget or constraint for the sampling allocation in
the surveillance design. Default is \code{NULL}. Units should be
consistent with \code{sample_cost} when specified. Otherwise the units
should be consistent with the \code{surv_qty_unit} parameter specified in
the \code{context} (e.g. traps or samples).}

\item{system_sens}{The desired (minimum) system sensitivity or detection
probability of the surveillance design (e.g. 0.95). Default is
\code{NULL}.}

\item{min_alloc}{A vector of minimum permissible sampling allocation at each
division part (location, category, etc.) specified by \code{divisions}.
Used to avoid impractically low sampling allocations. Default is
\code{NULL}.}

\item{max_alloc}{A vector of maximum permissible sampling allocation at each
division part (location, category, etc.) specified by \code{divisions}.
Used to avoid impractically high sampling allocations. Default is
\code{NULL}.}

\item{discrete_alloc}{A logical to indicate that the sampling allocation at
each division part (location, category, etc.) specified by
\code{divisions} should be discrete integers. Set to \code{TRUE} (default)
to allocate discrete samples. Set to \code{FALSE} for continuous
quantities if applicable within a sampling design, such as sample batches.}

\item{exist_alloc}{A vector of existing surveillance resource quantities at
each division part (location, category, etc.) specified by
\code{divisions}. Should only be used to represent existing surveillance
designs when \code{optimal = "none"}. Default is \code{NULL}.}

\item{exist_sens}{A vector, or list of vectors, of detection sensitivity
values of existing surveillance present at each division part (location,
category, etc.) specified by \code{divisions}. Multiple existing
surveillance layers may be specified in a list. Default is \code{NULL}.}

\item{...}{Additional parameters.}
}
\value{
A \code{SamplingSurvDesign} class object (list) containing inherited
and extended functions from the base \code{SurveillanceDesign} class for
for allocating samples, and calculating (unit and overall) detection
sensitivities:
\describe{
\item{\code{get_context()}}{Get context object.}
\item{\code{get_divisions()}}{Get divisions object.}
\item{\code{get_allocation()}}{Get allocated resources via specified
strategy, utilizing costs, benefits, budget constraints, and/or
desired system sensitivity or detection probability.}
\item{\code{get_sensitivity()}}{Get the division part detection
sensitivities of the allocated surveillance design combined with any
existing sensitivities specified via \code{exist_sens}.}
\item{\code{get_system_sens(growth = NULL)}}{Get the overall system
sensitivity or detection probability of the allocated surveillance
design. The optional \code{growth} parameter may provide a vector of
relative increasing multipliers (e.g. 1, 1.8, 4.3, 7.5) applied to the
discrete sampling design \code{prevalence}, or the continuous sampling
design density (\code{design_dens}), over time or a sequence of
repeated surveillance efforts, which provide a proxy for invasive
species growth. When present, increasing system sensitivity values are
returned for each multiplier or time/repeat.}
\item{\code{save_design(...)}}{Save the surveillance design as a
collection of raster TIF and/or comma-separated value (CSV) files,
appropriate for the \code{divisions} type, including the surveillance
\code{allocation}, the corresponding sensitivity \code{alloc_sens}
when existing sensitivity is included, the overall \code{sensitivity},
\code{surv_cost} (combined surveillance allocation and fixed costs),
and a \code{summary} (CSV) of the total allocation, total costs (when
applicable), and the overall allocation sensitivity
(\code{alloc_sens}) when applicable, and the system-wide sensitivity
(\code{system_sens}). \code{Terra} raster write options may be passed
to the function for saving grid-based designs.}
}
}
\description{
Builds a class to represent surveillance design functionality for the
effective allocation of surveillance sampling across one or more
divisions (parts, locations, categories, etc.) via Lagrange-based methods
for optimizing objective functions specified with surveillance and/or
incursion management costs, benefits, detection sensitivities, and/or an
overall desired system-wide sensitivity or detection probability.
}
\references{
Cannon, R. M. (2009). Inspecting and monitoring on a restricted budget -
where best to look? \emph{Preventive Veterinary Medicine}, 92(1–2),
163-174. \doi{10.1016/j.prevetmed.2009.06.009}

Kean, J. M., Burnip, G. M. & Pathan, A. (2015). Detection survey design
for decision making during biosecurity incursions. In F. Jarrad,
S. Low-Choy & K. Mengersen (eds.),
\emph{Biosecurity surveillance: Quantitative approaches} (pp. 238– 250).
Wallingford, UK: CABI.
}
