% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurveillanceDesign.R
\name{SurveillanceDesign}
\alias{SurveillanceDesign}
\alias{SurveillanceDesign.Raster}
\alias{SurveillanceDesign.SpatRaster}
\alias{SurveillanceDesign.Region}
\title{Generic surveillance design class builder}
\usage{
SurveillanceDesign(
  region,
  context,
  establish_pr,
  optimal = c("cost", "benefit", "detection"),
  mgmt_cost = NULL,
  benefit = NULL,
  alloc_units = NULL,
  fixed_cost = NULL,
  budget = NULL,
  confidence = NULL,
  exist_sens = NULL,
  ...
)

\method{SurveillanceDesign}{Raster}(region, ...)

\method{SurveillanceDesign}{SpatRaster}(region, ...)

\method{SurveillanceDesign}{Region}(
  region,
  context,
  establish_pr,
  optimal = c("cost", "benefit", "detection"),
  mgmt_cost = list(),
  benefit = NULL,
  alloc_units = NULL,
  fixed_cost = NULL,
  budget = NULL,
  confidence = NULL,
  exist_sens = NULL,
  ...
)
}
\arguments{
\item{region}{A \code{raster::RasterLayer}, \code{terra::SpatRaster}, or
\code{Region} or inherited class object representing the spatial region
(template) for the surveillance design.}

\item{context}{A \code{Context} or inherited class object representing the
context of a bio-security surveillance and area freedom design.}

\item{establish_pr}{A vector of (relative) probability values to represent
the likelihood of pest establishment at each location specified by the
\code{region}. Values are assumed to be relative when their range is
outside of 0-1, or an attribute \code{relative = TRUE} is attached to the
parameter.}

\item{optimal}{The strategy used for finding an effective surveillance
resource allocation. One of (minimum) \code{"cost"}, (maximum)
\code{"benefit"}, or (maximum) \code{"detection"} sensitivity (up to
\code{"confidence"} level when specified).}

\item{mgmt_cost}{A list of vectors to represent management costs specific to
the method implemented in the inherited class. Each vector specifies costs
at each location specified by the \code{region}. Default is empty list
(in the base class). An attribute \code{units} may be used to specify the
cost units (e.g. "$" or "hours").}

\item{benefit}{A vector of values quantifying the benefit of detection
at each location specified by the \code{region}. Default is \code{NULL}.
An attribute \code{units} may be used to specify the benefit units (e.g.
"$" or "hours").}

\item{alloc_units}{The units for the allocated surveillance resources (e.g.
"$", "hours", or "traps") consistent with the \code{context}. This may be
different to those specified for \code{mgmt_cost} or \code{benefit}.
Default is \code{NULL}.}

\item{fixed_cost}{A vector of fixed costs, such as travel costs, at each
location specified by the \code{region}. Default is \code{NULL}. Units are
specified in \code{alloc_units}.}

\item{budget}{The cost budget or constraint for the resource allocation in
the surveillance design. Default is \code{NULL}. Units are specified in
\code{alloc_units}.}

\item{confidence}{The desired (minimum) system detection sensitivity or
confidence of the surveillance design (e.g. 0.95). Default is \code{NULL}.}

\item{exist_sens}{A vector of detection sensitivity values of existing
surveillance present at each location specified by the \code{region}.
Default is \code{NULL}.}

\item{...}{Additional parameters.}
}
\value{
A \code{SurveillanceDesign} class object (list) containing functions
for allocating resources, and calculating (unit and overall) detection
sensitivities:
\describe{
\item{\code{get_allocation()}}{Get allocated surveillance resources via
specified strategy, utilizing costs, benefits, budget constraints,
and/or desired confidence level.}
\item{\code{get_sensitivity()}}{Get the unit/location detection
sensitivities of the allocated surveillance design.}
\item{\code{get_confidence()}}{Get the overall system sensitivity or
confidence of the allocated surveillance design.}
}
}
\description{
Builds a base class to represent a surveillance design with functionality
for the effective allocation of surveillance resources via methods that
consider surveillance and incursion management costs, benefits, detection
sensitivities, and/or overall detection confidence.
}
